refactor: SerachUser 컴포넌트에서 Input 컴포넌트 분리 및 이름 오타 수정
